+++ CHANGELOG.md
@@ -9,8 +9,7 @@
 
 ## 0.2.0
 ### ✨ Features and improvements
-
-### 🐞 Bug fixes
+- workflow test
 
 ## 0.1.0
 ### ✨ Features and improvements